project initialized and bug fixes in initialization + some optimizations todo: create other branches to add various features that are defined in header files

+++ gilli Danda/include/circularBufferSensorData.h
@@ -1,3 +1,8 @@
+/**
+ * @file circularBufferSensorData.h
+ * @brief This header file defines the CircularBuffer class and the SensorData structure used to manage the sensor data in a circular buffer format. The CircularBuffer class provides methods to insert new sensor data, retrieve the stored data in order, get the count of samples, and reset the buffer. The SensorData structure contains the timestamp and magnitude of the sensor readings.
+ */
+
 #ifndef CIRCULARBUFFERSENSORDATA_H
 #define CIRCULARBUFFERSENSORDATA_H
 #include "config.h"
@@ -12,7 +17,7 @@ class CircularBuffer
 private:
     SensorData dataBuffer[BUFFER_SIZE];
     SensorData res[BUFFER_SIZE];
-    int head;
+    size_t head;
     bool isFull;
 
 public:
@@ -25,7 +30,7 @@ public:
     SensorData* getData();
 
     // function to get total number of samples
-    int getSampleCount();
+    size_t getSampleCount();
 
     // function to reset buffer
     void resetBuffer();
